LayoutTests/http/tests/webxr/webxr-third-party-iframe-requestsession-denied-by-insufficient-feature-policy-expected.txt

CONSOLE MESSAGE: Feature policy 'XRSpatialTracking' check failed for element with origin 'http://localhost:8000' and allow attribute ''.
Tests that third-party iframes will be denied from requesting webxr session without xr-spatial-tracking feature policy.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS e.data.requestSession is "NotSupportedError"
PASS successfullyParsed is true

TEST COMPLETE
